pyproject.toml
[project]
name = "honk"
version = "0.1.0"
description = "Agent-first CLI for developer workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0,<0.12.0",
    "click>=8.0.0,<8.2.0",
    "rich==13.7.1",
    "textual==0.61.0",
    "httpx==0.27.2",
    "pydantic==2.8.2",
    "keyring==24.3.1",
    "jsonschema==4.23.0",
    "watchfiles>=1.1.1",
    "psutil>=7.1.3",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-xdist==3.6.1",
    "respx==0.21.1",
    "ruff==0.6.9",
    "mypy==1.11.2",
    "nox==2024.4.15",
]

[project.scripts]
honk = "honk.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/honk"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
